move convert dtype to functional_tensor since only works on tensors
